moved MviProfile model from app/models to lib/mvi/models
