+++ app/models/mvi.rb
@@ -73,7 +73,7 @@ class Mvi < Common::RedisStore
 
   # The profile returned from the MVI service. Either returned from cached response in Redis or the MVI service.
   #
-  # @return [MviProfile] patient 'golden record' data from MVI
+  # @return [MVI::Models::MviProfile] patient 'golden record' data from MVI
   def profile
     return nil unless @user.loa3?
     mvi_response&.profile
