+++ spec/factories/mvi_profiles.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 FactoryGirl.define do
-  factory :mvi_profile_address, class: 'MviProfileAddress' do
+  factory :mvi_profile_address, class: 'MVI::Models::MviProfileAddress' do
     street { Faker::Address.street_address }
     city { Faker::Address.city }
     state { Faker::Address.state }
@@ -24,7 +24,7 @@ FactoryGirl.define do
 end
 
 FactoryGirl.define do
-  factory :mvi_profile, class: 'MviProfile' do
+  factory :mvi_profile, class: 'MVI::Models::MviProfile' do
     given_names { Array.new(2) { Faker::Name.first_name } }
     family_name { Faker::Name.last_name }
     suffix { Faker::Name.suffix }
